@@ -1,22 +1,17 @@
-ext {
-    dashjVersion = "0.17.3"
-}
-
 buildscript {
-    ext.kotlin_version = '1.3.50'
-    ext.ok_http_version = "3.12.3"
-
+    ext.kotlin_version = '1.4.31'
+    ext.ok_http_version = '3.12.12'
+    ext.dashjVersion = "0.17.7"
     repositories {
         google()
+        mavenCentral()
         jcenter()
+        maven { url 'https://dl.bintray.com/kotlin/kotlin-eap' }
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.1'
-        classpath('fr.avianey.androidsvgdrawable:gradle-plugin:3.0.0') {
-            exclude group: 'xerces'
-        }
-        classpath 'com.google.gms:google-services:4.3.2'
+        classpath 'com.android.tools.build:gradle:4.1.3'
+        classpath 'com.google.gms:google-services:4.3.5'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -25,7 +20,8 @@ allprojects {
     repositories {
         google()
         mavenLocal()
+        mavenCentral()
+        maven { url 'https://jitpack.io' }
         jcenter()
-        maven { url 'https://oss.jfrog.org/artifactory/oss-snapshot-local' }
     }
 }
