@@ -1,11 +1,14 @@
 apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 29
+        vectorDrawables.useSupportLibrary = true
         versionCode 1
         versionName "1.0"
 
@@ -23,29 +26,45 @@ android {
             debuggable true
         }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'androidx.appcompat:appcompat:1.0.2'
-    api 'androidx.browser:browser:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    api 'androidx.browser:browser:1.3.0'
     implementation "org.dashj:dashj-core:$dashjVersion"
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
 
     //Retrofit + OkHttp
     api "com.squareup.okhttp3:okhttp:$ok_http_version"
     api "com.squareup.okhttp3:logging-interceptor:$ok_http_version"
-    api 'com.squareup.moshi:moshi:1.9.2'
+    api 'com.squareup.moshi:moshi:1.11.0'
     api 'com.squareup.retrofit2:retrofit:2.6.4'
     api 'com.squareup.retrofit2:converter-moshi:2.6.4'
     api 'com.squareup.retrofit2:converter-scalars:2.6.4'
+    api 'com.squareup.retrofit2:converter-gson:2.6.1'
 
     //Android Arch
-    api 'androidx.lifecycle:lifecycle-extensions:2.0.0'
-    api 'androidx.room:room-runtime:2.1.0'
-    annotationProcessor 'androidx.room:room-compiler:2.1.0'
+    api 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    api 'androidx.room:room-runtime:2.2.5'
+    annotationProcessor 'androidx.room:room-compiler:2.2.5'
 
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
+    testImplementation 'junit:junit:4.13.1'
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.3.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    implementation "androidx.core:core-ktx:+"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+}
+repositories {
+    mavenCentral()
 }
