@@ -85,8 +110,17 @@ ext {
     archivesBaseName = 'dash-wallet'
 }
 
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+
+def props = new Properties()
+if (file("../service.properties").exists()) {
+    file("../service.properties").withInputStream { props.load(it) }
+}
+
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
+    android.ndkVersion '21.3.6528147'
 
     lintOptions {
         abortOnError false
