@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2015 the original author or authors.
+ * Copyright the original author or authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -12,18 +12,23 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 package org.dash.wallet.common.ui;
 
-import android.annotation.SuppressLint;
+import androidx.appcompat.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface.OnClickListener;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AlertDialog;
+import android.graphics.drawable.Drawable;
 import android.view.LayoutInflater;
+import android.view.View;
+import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.DrawableRes;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
 
 import org.dash.wallet.common.R;
 
@@ -31,48 +36,119 @@ import org.dash.wallet.common.R;
  * @author Andreas Schildbach
  */
 public class DialogBuilder extends AlertDialog.Builder {
+    private final View customTitle;
+    private final ImageView iconView;
+    private final TextView titleView;
 
-    public static DialogBuilder warn(final Context context, final int titleResId) {
+    public static DialogBuilder dialog(final Context context, @StringRes final int titleResId,
+                                       @StringRes final int messageResId,
+                                       final Object... messageArgs) {
+        return dialog(context, titleResId, context.getString(messageResId, messageArgs));
+    }
+
+    public static DialogBuilder dialog(final Context context, @StringRes final int titleResId, final CharSequence message) {
         final DialogBuilder builder = new DialogBuilder(context);
-        builder.setIcon(R.drawable.ic_warning);
-        builder.setTitle(titleResId);
+        if (titleResId != 0)
+            builder.setTitle(titleResId);
+        builder.setMessage(message);
         return builder;
     }
 
+    public static DialogBuilder warn(final Context context, @StringRes final int titleResId,
+                                     @StringRes final int messageResId,
+                                     final Object... messageArgs) {
+        return warn(context, titleResId, context.getString(messageResId, messageArgs));
+    }
+
+    public static DialogBuilder warn(final Context context, @StringRes final int titleResId,
+                                     final CharSequence message) {
+        final DialogBuilder builder = dialog(context, titleResId, message);
+        builder.setIcon(R.drawable.ic_warning_grey600_24dp);
+        return builder;
+    }
+
+    public static DialogBuilder warn(final Context context, @StringRes final int titleResId) {
+        final DialogBuilder builder = new DialogBuilder(context);
+        builder.setTitle(titleResId);
+        builder.setIcon(R.drawable.ic_warning_grey600_24dp);
+        return builder;
+    }
+
+
+    public static DialogBuilder custom(final Context context, @StringRes final int titleResId, final View view) {
+        final DialogBuilder builder = new DialogBuilder(context);
+        if (titleResId != 0)
+            builder.setTitle(titleResId);
+        builder.setView(view);
+        return builder;
+    }
+
+    // make this protected
     public DialogBuilder(final Context context) {
-        super(context);
-    }
-
-    @SuppressLint("InflateParams")
-    @Override
-    public AlertDialog.Builder setTitle(int titleId) {
-        return this.setTitle(getContext().getString(titleId));
+        super(context, R.style.My_Theme_Dialog);
+        this.customTitle = LayoutInflater.from(context).inflate(R.layout.dialog_title, null);
+        this.iconView = customTitle.findViewById(android.R.id.icon);
+        this.titleView = customTitle.findViewById(android.R.id.title);
     }
 
     @Override
-    public AlertDialog.Builder setTitle(@androidx.annotation.Nullable CharSequence title) {
-        TextView titleTextView = (TextView) LayoutInflater.from(getContext())
-                .inflate(R.layout.dialog_title, null);
-        titleTextView.setText(title);
-        return this.setCustomTitle(titleTextView);
+    public DialogBuilder setIcon(final Drawable icon) {
+        if (icon != null) {
+            setCustomTitle(customTitle);
+            iconView.setImageDrawable(icon);
+            iconView.setVisibility(View.VISIBLE);
+        }
+
+        return this;
     }
 
     @Override
-    public DialogBuilder setMessage(final int messageResId) {
-        return this.setMessage(this.getContext().getString(messageResId));
+    public DialogBuilder setIcon(@DrawableRes final int iconResId) {
+        if (iconResId != 0) {
+            setCustomTitle(customTitle);
+            iconView.setImageResource(iconResId);
+            iconView.setVisibility(View.VISIBLE);
+        }
+
+        return this;
+    }
+
+    @Override
+    public DialogBuilder setTitle(final CharSequence title) {
+        if (title != null) {
+            setCustomTitle(customTitle);
+            titleView.setText(title);
+        }
+
+        return this;
+    }
+
+    @Override
+    public DialogBuilder setTitle(@StringRes final int titleResId) {
+        if (titleResId != 0) {
+            setCustomTitle(customTitle);
+            titleView.setText(titleResId);
+        }
+
+        return this;
     }
 
     @Override
     public DialogBuilder setMessage(final CharSequence message) {
-        TextView messageTextView = (TextView) LayoutInflater.from(getContext())
-                .inflate(R.layout.dialog_message, null);
-        messageTextView.setText(message);
-        this.setView(messageTextView);
+        super.setMessage(message);
+
+        return this;
+    }
+
+    @Override
+    public DialogBuilder setMessage(@StringRes final int messageResId) {
+        super.setMessage(messageResId);
+
         return this;
     }
 
     public DialogBuilder singleDismissButton(@Nullable final OnClickListener dismissListener) {
-        setNeutralButton(android.R.string.cancel, dismissListener);
+        setNeutralButton(R.string.button_dismiss, dismissListener);
 
         return this;
     }
