@@ -25,17 +25,20 @@ import retrofit2.http.Body;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
+import retrofit2.http.Header;
 import retrofit2.http.POST;
 import retrofit2.http.Path;
+import retrofit2.http.Url;
 
 public interface UpholdService {
 
     @FormUrlEncoded
     @POST("oauth2/token")
     Call<UpholdAccessToken> getAccessToken(@Field("client_id") String clientId,
-                                           @Field("client_secret")String clientSecret,
+                                           @Field("client_secret") String clientSecret,
                                            @Field("code") String code,
                                            @Field("grant_type") String grantType);
+
     @FormUrlEncoded
     @POST("oauth2/revoke")
     Call<String> revokeAccessToken(@Field("token") String token);
@@ -54,5 +57,9 @@ public interface UpholdService {
 
     @POST("v0/me/cards/{cardId}/transactions/{txId}/commit")
     Call<Object> commitTransaction(@Path("cardId") String cardId, @Path("txId") String txId);
+
+    @GET
+    Call<String> getUpholdCurrency(@Header("Range") String range, @Url String url);
+
 
 }
