@@ -21,17 +21,22 @@ import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.app.ProgressDialog;
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.net.Uri;
 import android.os.Bundle;
-import androidx.annotation.Nullable;
-import androidx.browser.customtabs.CustomTabsIntent;
-import androidx.core.content.ContextCompat;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.widget.Toolbar;
 import android.view.MenuItem;
 import android.view.View;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.widget.Toolbar;
+import androidx.browser.customtabs.CustomTabsIntent;
+import androidx.core.content.ContextCompat;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import org.dash.wallet.common.InteractionAwareActivity;
 import org.dash.wallet.common.customtabs.CustomTabActivityHelper;
@@ -39,10 +44,14 @@ import org.dash.wallet.integration.uphold.R;
 import org.dash.wallet.integration.uphold.data.UpholdClient;
 import org.dash.wallet.integration.uphold.data.UpholdConstants;
 
+
 public class UpholdSplashActivity extends InteractionAwareActivity {
+
+    public static final String FINISH_ACTION = "UpholdSplashActivity.FINISH_ACTION";
 
     public static final String UPHOLD_EXTRA_CODE = "uphold_extra_code";
     public static final String UPHOLD_EXTRA_STATE = "uphold_extra_state";
+    public static int taskId = 0;
 
     private ProgressDialog loadingDialog;
 
@@ -74,11 +83,20 @@ public class UpholdSplashActivity extends InteractionAwareActivity {
                 openLoginUrl();
             }
         });
+
+        handleIntent(getIntent());
+
+        IntentFilter filter = new IntentFilter(FINISH_ACTION);
+        LocalBroadcastManager.getInstance(this).registerReceiver(finishLinkReceiver, filter);
     }
 
     @Override
     protected void onNewIntent(Intent intent) {
         super.onNewIntent(intent);
+        handleIntent(intent);
+    }
+
+    private void handleIntent(final Intent intent) {
         Bundle extras = intent.getExtras();
         if (extras != null && extras.containsKey(UPHOLD_EXTRA_CODE)
                 && extras.containsKey(UPHOLD_EXTRA_STATE)) {
@@ -94,9 +112,17 @@ public class UpholdSplashActivity extends InteractionAwareActivity {
         overridePendingTransition(R.anim.activity_stay, R.anim.slide_out_left);
     }
 
+    private final BroadcastReceiver finishLinkReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            startActivity(intent); // this will ensure that the custom tab is closed
+        }
+    };
+
     @Override
     protected void onDestroy() {
         loadingDialog.dismiss();
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(finishLinkReceiver);
         super.onDestroy();
     }
 
@@ -152,6 +178,8 @@ public class UpholdSplashActivity extends InteractionAwareActivity {
     }
 
     private void openLoginUrl() {
+        //Intent intent = new Intent(this, CustomTabActivity.class);
+        //startActivity(intent);
         final String url = String.format(UpholdConstants.INITIAL_URL,
                 UpholdClient.getInstance().getEncryptionKey());
 
@@ -162,13 +190,14 @@ public class UpholdSplashActivity extends InteractionAwareActivity {
 
         CustomTabActivityHelper.openCustomTab(this, customTabsIntent, Uri.parse(url),
                 new CustomTabActivityHelper.CustomTabFallback() {
-            @Override
-            public void openUri(Activity activity, Uri uri) {
-                Intent intent = new Intent(Intent.ACTION_VIEW);
-                intent.setData(Uri.parse(url));
-                startActivity(intent);
-            }
-        });
+                    @Override
+                    public void openUri(Activity activity, Uri uri) {
+                        Intent intent = new Intent(Intent.ACTION_VIEW);
+                        intent.setData(Uri.parse(url));
+                        startActivity(intent);
+                    }
+                });
+        //finish();
     }
 
     @Override
