@@ -4,17 +4,19 @@ import android.app.Activity;
 import android.app.Application;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
+
 public abstract class ActivitiesTracker implements Application.ActivityLifecycleCallbacks {
 
     private int numStarted;
 
     @Override
-    public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
+    public void onActivityCreated(@NonNull Activity activity, Bundle savedInstanceState) {
 
     }
 
     @Override
-    public void onActivityStarted(Activity activity) {
+    public void onActivityStarted(@NonNull Activity activity) {
         boolean isTheFirstOne = numStarted == 0;
         if (isTheFirstOne) {
             onStartedFirst(activity);
@@ -24,17 +26,17 @@ public abstract class ActivitiesTracker implements Application.ActivityLifecycle
     }
 
     @Override
-    public void onActivityResumed(Activity activity) {
+    public void onActivityResumed(@NonNull Activity activity) {
 
     }
 
     @Override
-    public void onActivityPaused(Activity activity) {
+    public void onActivityPaused(@NonNull Activity activity) {
 
     }
 
     @Override
-    public void onActivityStopped(Activity activity) {
+    public void onActivityStopped(@NonNull Activity activity) {
         numStarted--;
         if (numStarted == 0) {
             // app went to background
@@ -43,12 +45,13 @@ public abstract class ActivitiesTracker implements Application.ActivityLifecycle
     }
 
     @Override
-    public void onActivitySaveInstanceState(Activity activity, Bundle outState) {
+    public void onActivitySaveInstanceState(@NonNull Activity activity,
+                                            @NonNull Bundle outState) {
 
     }
 
     @Override
-    public void onActivityDestroyed(Activity activity) {
+    public void onActivityDestroyed(@NonNull Activity activity) {
 
     }
 
