@@ -16,10 +16,17 @@
 
 package de.schildbach.wallet;
 
+import android.app.KeyguardManager;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.os.Handler;
 
 import org.dash.wallet.common.Configuration;
+import org.dash.wallet.common.InteractionAwareActivity;
 
 import java.util.concurrent.TimeUnit;
 
@@ -27,27 +34,31 @@ public class AutoLogout {
 
     private static final long LOCK_TIMER_TICK_MS = TimeUnit.SECONDS.toMillis(5);
 
-    private Handler lockTimerClock = new Handler();
+    private final Handler lockTimerClock = new Handler();
     private long tickCounter;
     private boolean timerActive = false;
 
-    private Configuration config;
+    private final Configuration config;
 
-    private boolean appInBackground = false;
+    private boolean appWentBackground = true;
+    public boolean keepLockedUntilPinEntered = true;
 
     private OnLogoutListener onLogoutListener;
+
+    public boolean deviceWasLocked = false;
 
     @SuppressWarnings("FieldCanBeLocal")
     private SharedPreferences.OnSharedPreferenceChangeListener configListener = new SharedPreferences.OnSharedPreferenceChangeListener() {
         @Override
         public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
             if (Configuration.PREFS_KEY_AUTO_LOGOUT_ENABLED.equals(key) || Configuration.PREFS_KEY_AUTO_LOGOUT_MINUTES.equals(key)) {
+                setAppWentBackground(false);
                 setup();
             }
         }
     };
 
-    AutoLogout(Configuration config) {
+    public AutoLogout(Configuration config) {
         this.config = config;
         this.config.registerOnSharedPreferenceChangeListener(configListener);
     }
@@ -63,18 +74,23 @@ public class AutoLogout {
         }
     }
 
+    public void maybeStartAutoLogoutTimer() {
+        setup();
+    }
+
     public void startTimer() {
         lockTimerClock.postDelayed(timerTask, LOCK_TIMER_TICK_MS);
         timerActive = true;
     }
 
-    private Runnable timerTask = new Runnable() {
+    private final Runnable timerTask = new Runnable() {
         @Override
         public void run() {
             tickCounter += LOCK_TIMER_TICK_MS;
             if (shouldLogout()) {
                 if (onLogoutListener != null) {
-                    onLogoutListener.onLogout(appInBackground);
+                    onLogoutListener.onLogout(appWentBackground);
+                    WalletApplication.getInstance().sendBroadcast(new Intent(InteractionAwareActivity.FORCE_FINISH_ACTION));
                 }
                 timerActive = false;
             } else {
@@ -85,8 +101,8 @@ public class AutoLogout {
 
     public boolean shouldLogout() {
         long autoLogoutMillis = TimeUnit.MINUTES.toMillis(config.getAutoLogoutMinutes());
-        boolean logoutTimeExceeded = (config.getAutoLogoutMinutes() == 0) || tickCounter >= autoLogoutMillis;
-        return config.getAutoLogoutEnabled() && logoutTimeExceeded;
+        boolean logoutTimeExceeded = (config.getAutoLogoutMinutes() == 0) ? appWentBackground : (tickCounter >= autoLogoutMillis);
+        return (config.getAutoLogoutEnabled() && logoutTimeExceeded) || deviceWasLocked;
     }
 
     public void stopTimer() {
@@ -109,15 +125,28 @@ public class AutoLogout {
         }
     }
 
-    public void setAppInBackground(boolean appInBackground) {
-        this.appInBackground = appInBackground;
+    public void setAppWentBackground(boolean appWentBackground) {
+        this.appWentBackground = appWentBackground;
     }
 
     public boolean isTimerActive() {
         return timerActive;
     }
 
-    interface OnLogoutListener {
+    public void registerDeviceInteractiveReceiver(Context context) {
+        final IntentFilter filter = new IntentFilter();
+        filter.addAction(Intent.ACTION_SCREEN_ON);
+        filter.addAction(Intent.ACTION_SCREEN_OFF);
+        context.registerReceiver(new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, Intent intent) {
+                KeyguardManager myKM = (KeyguardManager) context.getSystemService(Context.KEYGUARD_SERVICE);
+                deviceWasLocked |= Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP_MR1 ? myKM.isDeviceLocked() : myKM.inKeyguardRestrictedInputMode();
+            }
+        }, filter);
+    }
+
+    public interface OnLogoutListener {
         void onLogout(boolean isAppInBackground);
     }
 }
