@@ -155,7 +155,7 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
     private final SeedPeers seedPeerDiscovery = new SeedPeers(Constants.NETWORK_PARAMETERS);
     private final DnsDiscovery dnsDiscovery = new DnsDiscovery(Constants.DNS_SEED, Constants.NETWORK_PARAMETERS);
     ArrayList<PeerDiscovery> peerDiscoveryList = new ArrayList<>(2);
-
+    private final static int MINIMUM_PEER_COUNT = 16;
 
     private static final int MIN_COLLECT_HISTORY = 2;
     private static final int IDLE_BLOCK_TIMEOUT_MIN = 2;
@@ -518,34 +518,36 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
                         }
 
                         if (!connectTrustedPeerOnly) {
+                            // First use the masternode list that is included
                             try {
-                                peers.addAll(
-                                        Arrays.asList(normalPeerDiscovery.getPeers(services, timeoutValue, timeoutUnit)));
+                                SimplifiedMasternodeList mnlist = org.bitcoinj.core.Context.get().masternodeListManager.getListAtChainTip();
+                                MasternodePeerDiscovery discovery = new MasternodePeerDiscovery(mnlist);
+                                peers.addAll(Arrays.asList(discovery.getPeers(services, timeoutValue, timeoutUnit)));
                             } catch (PeerDiscoveryException x) {
-                                //swallow and continue with another method of connection.
-                                log.info("DNS peer discovery failed: "+ x.getMessage());
-                                if(x.getCause() != null)
-                                    log.info(  "cause:  " + x.getCause().getMessage());
+                                //swallow and continue with another method of connection
+                                log.info("DMN List peer discovery failed: "+ x.getMessage());
                             }
-                            if(peers.size() < 10) {
-                                log.info("DNS peer discovery returned less than 10 nodes.  Adding DMN peers to the list to increase connections");
-                                try {
-                                    SimplifiedMasternodeList mnlist = org.bitcoinj.core.Context.get().masternodeListManager.getListAtChainTip();
-                                    MasternodePeerDiscovery discovery = new MasternodePeerDiscovery(mnlist);
-                                    peers.addAll(Arrays.asList(discovery.getPeers(services, timeoutValue, timeoutUnit)));
-                                } catch (PeerDiscoveryException x) {
-                                    //swallow and continue with another method of connection
-                                    log.info("DMN List peer discovery failed: "+ x.getMessage());
 
+                            if(peers.size() < MINIMUM_PEER_COUNT) {
+                                if (Constants.NETWORK_PARAMETERS.getAddrSeeds() != null) {
+                                    log.info("DNM peer discovery returned less than 16 nodes.  Adding seed peers to the list to increase connections");
+                                    peers.addAll(Arrays.asList(seedPeerDiscovery.getPeers(services, timeoutValue, timeoutUnit)));
+                                } else {
+                                    log.info("DNS peer discovery returned less than 16 nodes.  Unable to add seed peers (it is not specified for this network).");
                                 }
+                            }
 
-                                if(peers.size() < 10) {
-                                    if (Constants.NETWORK_PARAMETERS.getAddrSeeds() != null) {
-                                        log.info("DNS peer discovery returned less than 10 nodes.  Adding seed peers to the list to increase connections");
-                                        peers.addAll(Arrays.asList(seedPeerDiscovery.getPeers(services, timeoutValue, timeoutUnit)));
-                                    } else {
-                                        log.info("DNS peer discovery returned less than 10 nodes.  Unable to add seed peers (it is not specified for this network).");
-                                    }
+                            if(peers.size() < MINIMUM_PEER_COUNT) {
+                                log.info("Masternode peer discovery returned less than 16 nodes.  Adding DMN peers to the list to increase connections");
+
+                                try {
+                                    peers.addAll(
+                                            Arrays.asList(normalPeerDiscovery.getPeers(services, timeoutValue, timeoutUnit)));
+                                } catch (PeerDiscoveryException x) {
+                                    //swallow and continue with another method of connection, if one exists.
+                                    log.info("DNS peer discovery failed: "+ x.getMessage());
+                                    if(x.getCause() != null)
+                                        log.info(  "cause:  " + x.getCause().getMessage());
                                 }
                             }
                         }
@@ -683,9 +685,6 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
         super.onCreate();
 
         application = (WalletApplication) getApplication();
-        if (application.getWallet() == null) {
-            return;
-        }
 
         nm = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
         connectivityManager = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
@@ -792,11 +791,6 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
     @Override
     public int onStartCommand(final Intent intent, final int flags, final int startId) {
         super.onStartCommand(intent, flags, startId);
-
-        if (application.getWallet() == null) {
-            log.warn("service started before wallet initialization");
-            return START_NOT_STICKY;
-        }
 
         if (intent != null) {
             //Restart service as a Foreground Service if it's synchronizing the blockchain
