@@ -4,10 +4,9 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import de.schildbach.wallet.livedata.Status
 import de.schildbach.wallet.ui.preference.PinRetryController
-import de.schildbach.wallet.util.FingerprintHelper
 import kotlinx.android.synthetic.main.fragment_enter_pin.*
 import org.bitcoinj.wallet.DeterministicSeed
 
@@ -50,13 +49,13 @@ class DecryptSeedWithPinDialog : CheckPinDialog() {
     }
 
     override fun initViewModel() {
-        viewModel = ViewModelProviders.of(this).get(DecryptSeedViewModel::class.java)
+        viewModel = ViewModelProvider(this)[DecryptSeedViewModel::class.java]
         (viewModel as DecryptSeedViewModel).decryptSeedLiveData.observe(viewLifecycleOwner, Observer {
             when (it.status) {
                 Status.ERROR -> {
                     pinRetryController.failedAttempt(it.data!!.second!!)
                     if (pinRetryController.isLocked) {
-                        showLockedAlert(context!!)
+                        showLockedAlert(requireContext())
                         dismiss()
                         return@Observer
                     }
@@ -76,14 +75,14 @@ class DecryptSeedWithPinDialog : CheckPinDialog() {
         if (pinRetryController.isLocked) {
             return
         }
-        val requestCode = arguments!!.getInt(ARG_REQUEST_CODE)
+        val requestCode = requireArguments().getInt(ARG_REQUEST_CODE)
         (sharedModel as DecryptSeedSharedModel).onDecryptSeedCallback.value = Pair(requestCode, seed)
         pinRetryController.clearPinFailPrefs()
         dismiss()
     }
 
     override fun FragmentActivity.initSharedModel(activity: FragmentActivity) {
-        sharedModel = ViewModelProviders.of(activity)[DecryptSeedSharedModel::class.java]
+        sharedModel = ViewModelProvider(activity)[DecryptSeedSharedModel::class.java]
     }
 
     override fun onFingerprintSuccess(savedPass : String) {
