@@ -20,11 +20,10 @@ import android.annotation.SuppressLint
 import android.content.Intent
 import de.schildbach.wallet.ui.widget.GlobalFooterView
 import de.schildbach.wallet_test.R
-import org.dash.wallet.common.InteractionAwareActivity
 
 
 @SuppressLint("Registered")
-open class GlobalFooterActivity : InteractionAwareActivity(), GlobalFooterView.OnFooterActionListener {
+open class GlobalFooterActivity : LockScreenActivity(), GlobalFooterView.OnFooterActionListener {
 
     private lateinit var globalFooterView: GlobalFooterView
 
