@@ -17,6 +17,7 @@
 package de.schildbach.wallet.ui
 
 import android.annotation.SuppressLint
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.graphics.drawable.LayerDrawable
@@ -30,6 +31,7 @@ import android.widget.Toast
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import de.schildbach.wallet.WalletApplication
+import de.schildbach.wallet.ui.backup.RestoreFromFileActivity
 import de.schildbach.wallet.ui.preference.PinRetryController
 import de.schildbach.wallet.ui.security.SecurityGuard
 import de.schildbach.wallet_test.R
@@ -38,6 +40,9 @@ import kotlinx.android.synthetic.main.activity_onboarding_perm_lock.*
 import org.dash.wallet.common.ui.DialogBuilder
 
 private const val REGULAR_FLOW_TUTORIAL_REQUEST_CODE = 0
+const val SET_PIN_REQUEST_CODE = 1
+private const val RESTORE_PHRASE_REQUEST_CODE = 2
+private const val RESTORE_FILE_REQUEST_CODE = 3
 
 class OnboardingActivity : RestoreFromFileActivity() {
 
@@ -45,13 +50,16 @@ class OnboardingActivity : RestoreFromFileActivity() {
         @JvmStatic
         fun createIntent(context: Context): Intent {
             return Intent(context, OnboardingActivity::class.java)
-                    .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
         }
     }
 
     private lateinit var viewModel: OnboardingViewModel
 
     private lateinit var walletApplication: WalletApplication
+
+    override fun onStart() {
+        super.onStart()
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -75,6 +83,7 @@ class OnboardingActivity : RestoreFromFileActivity() {
         viewModel = ViewModelProviders.of(this).get(OnboardingViewModel::class.java)
 
         walletApplication = (application as WalletApplication)
+
         if (walletApplication.walletFileExists()) {
             regularFlow()
         } else {
@@ -110,12 +119,7 @@ class OnboardingActivity : RestoreFromFileActivity() {
     }
 
     private fun startMainActivity() {
-        val intent = if (walletApplication.configuration.autoLogoutEnabled) {
-            LockScreenActivity.createIntent(this)
-        } else {
-            WalletActivity.createIntent(this)
-        }
-        startActivity(intent)
+        startActivity(WalletActivity.createIntent(this))
         finish()
     }
 
@@ -135,7 +139,7 @@ class OnboardingActivity : RestoreFromFileActivity() {
         }
         recovery_wallet.setOnClickListener {
             walletApplication.initEnvironmentIfNeeded()
-            startActivity(Intent(this, RestoreWalletFromSeedActivity::class.java))
+            startActivityForResult(Intent(this, RestoreWalletFromSeedActivity::class.java), REQUEST_CODE_RESTORE_WALLET)
         }
         restore_wallet.setOnClickListener {
             restoreWalletFromFile()
@@ -161,7 +165,7 @@ class OnboardingActivity : RestoreFromFileActivity() {
             dialog.show()
         })
         viewModel.startActivityAction.observe(this, Observer {
-            startActivity(it)
+            startActivityForResult(it, SET_PIN_REQUEST_CODE)
         })
     }
 
@@ -197,6 +201,12 @@ class OnboardingActivity : RestoreFromFileActivity() {
         super.onActivityResult(requestCode, resultCode, data)
         if (requestCode == REGULAR_FLOW_TUTORIAL_REQUEST_CODE) {
             upgradeOrStartMainActivity()
+        } else if ((requestCode == SET_PIN_REQUEST_CODE || requestCode == RESTORE_PHRASE_REQUEST_CODE) && resultCode == Activity.RESULT_OK) {
+            finish()
         }
+    }
+
+    fun getWalletApplication() : WalletApplication {
+        return walletApplication
     }
 }
