@@ -16,6 +16,7 @@
 
 package de.schildbach.wallet.ui
 
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
@@ -26,7 +27,7 @@ import android.widget.TextView
 import android.widget.ViewSwitcher
 import androidx.appcompat.widget.Toolbar
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import de.schildbach.wallet.WalletApplication
 import de.schildbach.wallet.livedata.Status
 import de.schildbach.wallet.ui.preference.PinRetryController
@@ -42,7 +43,7 @@ class SetPinActivity : InteractionAwareActivity() {
     private lateinit var numericKeyboardView: NumericKeyboardView
     private lateinit var confirmButtonView: View
     private lateinit var viewModel: SetPinViewModel
-    private lateinit var enableFingerprintViewModel: CheckPinSharedModel
+    private lateinit var enableFingerprintViewModel: EnableFingerprintDialog.SharedViewModel
     private lateinit var pinProgressSwitcherView: ViewSwitcher
     private lateinit var pinPreviewView: PinPreviewView
     private lateinit var pageTitleView: TextView
@@ -54,7 +55,13 @@ class SetPinActivity : InteractionAwareActivity() {
     val pin = arrayListOf<Int>()
     var seed = listOf<String>()
 
-    private var changePin = false
+    private val initialPin by lazy {
+        intent.getStringExtra(EXTRA_PASSWORD)
+    }
+
+    private val changePin by lazy {
+        intent.getBooleanExtra(CHANGE_PIN, false)
+    }
 
     private enum class State {
         DECRYPT,
@@ -78,11 +85,11 @@ class SetPinActivity : InteractionAwareActivity() {
         @JvmOverloads
         @JvmStatic
         fun createIntent(context: Context, titleResId: Int,
-                         changePin: Boolean = false, password: String? = null): Intent {
+                         changePin: Boolean = false, pin: String? = null): Intent {
             val intent = Intent(context, SetPinActivity::class.java)
             intent.putExtra(EXTRA_TITLE_RES_ID, titleResId)
             intent.putExtra(CHANGE_PIN, changePin)
-            intent.putExtra(EXTRA_PASSWORD, password)
+            intent.putExtra(EXTRA_PASSWORD, pin)
             return intent
         }
 
@@ -109,10 +116,13 @@ class SetPinActivity : InteractionAwareActivity() {
 
         walletApplication = application as WalletApplication
         if (walletApplication.wallet.isEncrypted) {
-            val password = intent.getStringExtra(EXTRA_PASSWORD)
-            changePin = intent.getBooleanExtra(CHANGE_PIN, false)
-            if (password != null) {
-                viewModel.decryptKeys(password)
+            if (initialPin != null) {
+                if (changePin) {
+                    viewModel.oldPinCache = initialPin
+                    setState(State.SET_PIN)
+                } else {
+                    viewModel.decryptKeys(initialPin)
+                }
             } else {
                 if (changePin) {
                     if (pinRetryController.isLocked) {
@@ -316,7 +326,7 @@ class SetPinActivity : InteractionAwareActivity() {
     }
 
     private fun initViewModel() {
-        viewModel = ViewModelProviders.of(this).get(SetPinViewModel::class.java)
+        viewModel = ViewModelProvider(this)[SetPinViewModel::class.java]
         viewModel.encryptWalletLiveData.observe(this, Observer {
             when (it.status) {
                 Status.ERROR -> {
@@ -328,7 +338,7 @@ class SetPinActivity : InteractionAwareActivity() {
                     } else {
                         if (state == State.DECRYPTING) {
                             setState(if (changePin) State.INVALID_PIN else State.DECRYPT)
-                            if(!changePin) {
+                            if (!changePin) {
                                 android.widget.Toast.makeText(this, "Incorrect PIN", android.widget.Toast.LENGTH_LONG).show()
                             }
                         } else {
@@ -351,7 +361,12 @@ class SetPinActivity : InteractionAwareActivity() {
                             if (EnableFingerprintDialog.shouldBeShown(this@SetPinActivity) && enableFingerprint) {
                                 EnableFingerprintDialog.show(viewModel.getPinAsString(), supportFragmentManager)
                             } else {
-                                finish()
+                                if (initialPin != null) {
+                                    pinRetryController.clearPinFailPrefs()
+                                    goHome()
+                                } else {
+                                    finish()
+                                }
                             }
                         } else {
                             viewModel.initWallet()
@@ -381,27 +396,25 @@ class SetPinActivity : InteractionAwareActivity() {
             }
         })
         viewModel.startNextActivity.observe(this, Observer {
+            setResult(Activity.RESULT_OK)
             if (it) {
                 startVerifySeedActivity()
             } else {
                 goHome()
             }
-            walletApplication.maybeStartAutoLogoutTimer()
+            walletApplication.autoLogout.apply {
+                maybeStartAutoLogoutTimer()
+                keepLockedUntilPinEntered = false
+            }
         })
-        enableFingerprintViewModel = ViewModelProviders.of(this).get(CheckPinSharedModel::class.java)
+        enableFingerprintViewModel = ViewModelProvider(this)[EnableFingerprintDialog.SharedViewModel::class.java]
         enableFingerprintViewModel.onCorrectPinCallback.observe(this, Observer {
-            finish()
+            if (initialPin != null) {
+                goHome()
+            } else {
+                finish()
+            }
         })
-    }
-
-    private fun startActivityNewTask(intent: Intent) {
-        intent.apply {
-            addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-            addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
-        startActivity(intent)
-        finish()
     }
 
     override fun finish() {
@@ -431,12 +444,12 @@ class SetPinActivity : InteractionAwareActivity() {
     }
 
     private fun startVerifySeedActivity() {
-        val intent = VerifySeedActivity.createIntent(this, seed.toTypedArray())
-        startActivityNewTask(intent)
+        startActivity(VerifySeedActivity.createIntent(this, seed.toTypedArray()))
+        finish()
     }
 
     private fun goHome() {
-        val intent = Intent(this, WalletActivity::class.java)
-        startActivityNewTask(intent)
+        startActivity(WalletActivity.createIntent(this))
+        finish()
     }
 }
