@@ -34,26 +34,18 @@ import android.nfc.NdefMessage;
 import android.nfc.NfcAdapter;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
 import android.os.Handler;
 import android.os.LocaleList;
 import android.telephony.TelephonyManager;
-import android.view.ContextMenu;
 import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.widget.Toolbar;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
@@ -68,9 +60,9 @@ import org.bitcoinj.core.VerificationException;
 import org.bitcoinj.crypto.ChildNumber;
 import org.bitcoinj.wallet.Wallet;
 import org.dash.wallet.common.Configuration;
+import org.dash.wallet.common.UserInteractionAwareCallback;
 import org.dash.wallet.common.data.CurrencyInfo;
 import org.dash.wallet.common.ui.DialogBuilder;
-import org.dash.wallet.integration.uphold.ui.UpholdAccountActivity;
 
 import java.io.IOException;
 import java.util.Currency;
@@ -83,13 +75,15 @@ import de.schildbach.wallet.WalletBalanceWidgetProvider;
 import de.schildbach.wallet.data.PaymentIntent;
 import de.schildbach.wallet.ui.InputParser.BinaryInputParser;
 import de.schildbach.wallet.ui.InputParser.StringInputParser;
+import de.schildbach.wallet.ui.backup.BackupWalletDialogFragment;
+import de.schildbach.wallet.ui.backup.RestoreFromFileHelper;
 import de.schildbach.wallet.ui.preference.PreferenceActivity;
 import de.schildbach.wallet.ui.scan.ScanActivity;
 import de.schildbach.wallet.ui.send.SendCoinsInternalActivity;
 import de.schildbach.wallet.ui.send.SweepWalletActivity;
+import de.schildbach.wallet.ui.widget.ShortcutsPane;
 import de.schildbach.wallet.ui.widget.UpgradeWalletDisclaimerDialog;
 import de.schildbach.wallet.util.CrashReporter;
-import de.schildbach.wallet.util.FingerprintHelper;
 import de.schildbach.wallet.util.Nfc;
 import de.schildbach.wallet_test.R;
 import kotlin.Pair;
@@ -117,11 +111,8 @@ public final class WalletActivity extends AbstractBindServiceActivity
     private WalletApplication application;
     private Configuration config;
     private Wallet wallet;
-    private FingerprintHelper fingerprintHelper;
 
-    private DrawerLayout viewDrawer;
-    private View viewFakeForSafetySubmenu;
-    private View payBtn;
+    private ShortcutsPane shortcutsPane;
 
     private Handler handler = new Handler();
 
@@ -145,7 +136,7 @@ public final class WalletActivity extends AbstractBindServiceActivity
         wallet = application.getWallet();
 
         setContentViewFooter(R.layout.home_activity);
-        setSupportActionBar((Toolbar) findViewById(R.id.toolbar));
+        setSupportActionBar(findViewById(R.id.toolbar));
         activateHomeButton();
 
         if (savedInstanceState == null) {
@@ -165,8 +156,6 @@ public final class WalletActivity extends AbstractBindServiceActivity
             //Add BIP44 support and PIN if missing
             upgradeWalletKeyChains(Constants.BIP44_PATH, false);
         }
-
-        initFingerprintHelper();
 
         this.clipboardManager = (ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
 
@@ -194,97 +183,40 @@ public final class WalletActivity extends AbstractBindServiceActivity
         });
     }
 
-    private void initFingerprintHelper() {
-        //Init fingerprint helper
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            fingerprintHelper = new FingerprintHelper(this);
-            if (!fingerprintHelper.init()) {
-                fingerprintHelper = null;
-            }
-        }
-    }
-
     private void initView() {
-        initNavigationDrawer();
-        initQuickActions();
-        findViewById(R.id.uphold_account_section).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startUpholdActivity();
-                viewDrawer.closeDrawer(GravityCompat.START);
-            }
-        });
-        findViewById(R.id.pay_btn).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(PaymentsActivity.createIntent(WalletActivity.this, PaymentsActivity.ACTIVE_TAB_PAY));
-            }
-        });
-        findViewById(R.id.receive_btn).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(PaymentsActivity.createIntent(WalletActivity.this, PaymentsActivity.ACTIVE_TAB_RECEIVE));
-            }
-        });
+        initShortcutActions();
     }
 
-    private void initQuickActions() {
+    private void initShortcutActions() {
+        shortcutsPane = findViewById(R.id.shortcuts_pane);
+        shortcutsPane.setOnShortcutClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (v == shortcutsPane.getSecureNowButton()) {
+                    handleBackupWalletToSeed();
+                } else if (v == shortcutsPane.getScanToPayButton()) {
+                    handleScan(v);
+                } else if (v == shortcutsPane.getBuySellButton()) {
+                    startUpholdActivity();
+                } else if (v == shortcutsPane.getPayToAddressButton()) {
+                    handlePaste();
+                } else if (v == shortcutsPane.getReceiveButton()) {
+                    startActivity(PaymentsActivity.createIntent(WalletActivity.this, PaymentsActivity.ACTIVE_TAB_RECEIVE));
+                } else if (v == shortcutsPane.getImportPrivateKey()) {
+                    SweepWalletActivity.start(WalletActivity.this, true);
+                }
+            }
+        });
         showHideSecureAction();
-        findViewById(R.id.secure_action).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                handleBackupWalletToSeed();
-            }
-        });
-        findViewById(R.id.scan_to_pay_action).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                handleScan(v);
-            }
-        });
-        findViewById(R.id.buy_sell_action).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startUpholdActivity();
-            }
-        });
-        findViewById(R.id.pay_to_address_action).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                handlePaste();
-            }
-        });
-        findViewById(R.id.import_key_action).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                SweepWalletActivity.start(WalletActivity.this, true);
-            }
-        });
     }
 
     private void showHideSecureAction() {
-        View secureActionView = findViewById(R.id.secure_action);
-        secureActionView.setVisibility(config.getRemindBackupSeed() ? View.VISIBLE : View.GONE);
-        findViewById(R.id.secure_action_space).setVisibility(secureActionView.getVisibility());
-    }
-
-    private void initNavigationDrawer() {
-        final NavigationView navigationView = findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
-
-        viewFakeForSafetySubmenu = new View(this);
-        viewFakeForSafetySubmenu.setVisibility(View.GONE);
-        viewDrawer = findViewById(R.id.drawer_layout);
-        viewDrawer.addView(viewFakeForSafetySubmenu);
-        registerForContextMenu(viewFakeForSafetySubmenu);
-        viewDrawer.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
+        shortcutsPane.showSecureNow(config.getRemindBackupSeed());
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-
-        getWalletApplication().startBlockchainService(true);
 
         checkLowStorageAlert();
         detectUserCountry();
@@ -530,7 +462,8 @@ public final class WalletActivity extends AbstractBindServiceActivity
     }
 
     private void handleReportIssue() {
-        ReportIssueDialogBuilder.createReportIssueDialog(this, application).show();
+        Dialog dialog = ReportIssueDialogBuilder.createReportIssueDialog(this, application).show();
+        dialog.getWindow().setCallback(new UserInteractionAwareCallback(dialog.getWindow().getCallback(), this));
     }
 
     private void handlePaste() {
@@ -572,8 +505,8 @@ public final class WalletActivity extends AbstractBindServiceActivity
             return createBackupWalletPermissionDialog();
         else if (id == DIALOG_RESTORE_WALLET_PERMISSION)
             return createRestoreWalletPermissionDialog();
-        else if (id == DIALOG_RESTORE_WALLET)
-            return createRestoreWalletDialog();
+        //else if (id == DIALOG_RESTORE_WALLET)
+        //    return createRestoreWalletDialog();
         else if (id == DIALOG_TIMESKEW_ALERT)
             return createTimeskewAlertDialog(args.getLong("diff_minutes"));
         else if (id == DIALOG_VERSION_ALERT)
@@ -584,11 +517,11 @@ public final class WalletActivity extends AbstractBindServiceActivity
             throw new IllegalArgumentException();
     }
 
-    @Override
+    /*@Override
     protected void onPrepareDialog(final int id, final Dialog dialog) {
         if (id == DIALOG_RESTORE_WALLET)
             prepareRestoreWalletDialog(dialog);
-    }
+    }*/
 
     private Dialog createBackupWalletPermissionDialog() {
         final DialogBuilder dialog = new DialogBuilder(this);
@@ -600,26 +533,6 @@ public final class WalletActivity extends AbstractBindServiceActivity
 
     private Dialog createRestoreWalletPermissionDialog() {
         return RestoreFromFileHelper.createRestoreWalletPermissionDialog(this);
-    }
-
-    private Dialog createRestoreWalletDialog() {
-        return RestoreFromFileHelper.createRestoreWalletDialog(this, new RestoreFromFileHelper.OnRestoreWalletListener() {
-            @Override
-            public void onRestoreWallet(Wallet wallet) {
-                restoreWallet(wallet);
-                application.getConfiguration().setRestoringBackup(true);
-            }
-
-            @Override
-            public void onRetryRequest() {
-                showDialog(DIALOG_RESTORE_WALLET);
-            }
-        });
-    }
-
-    private void prepareRestoreWalletDialog(final Dialog dialog) {
-        final boolean hasCoins = wallet.getBalance(Wallet.BalanceType.ESTIMATED).signum() > 0;
-        RestoreFromFileHelper.prepareRestoreWalletDialog(this, hasCoins, dialog);
     }
 
     private void showRestoreWalletFromSeedDialog() {
@@ -827,10 +740,6 @@ public final class WalletActivity extends AbstractBindServiceActivity
         config.disarmBackupReminder();
         this.wallet = application.getWallet();
         upgradeWalletKeyChains(Constants.BIP44_PATH, true);
-
-        if (fingerprintHelper != null) {
-            fingerprintHelper.clear();
-        }
     }
 
     private void resetBlockchain() {
@@ -859,26 +768,6 @@ public final class WalletActivity extends AbstractBindServiceActivity
             handleEncryptKeysRestoredWallet();
         } else {
             resetBlockchain();
-        }
-    }
-
-    @Override
-    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
-        super.onCreateContextMenu(menu, v, menuInfo);
-        MenuInflater inflater = getMenuInflater();
-        if (v == viewFakeForSafetySubmenu) {
-            inflater.inflate(R.menu.wallet_safety_options, menu);
-
-            final String externalStorageState = Environment.getExternalStorageState();
-
-            menu.findItem(R.id.wallet_options_restore_wallet).setEnabled(
-                    Environment.MEDIA_MOUNTED.equals(externalStorageState) || Environment.MEDIA_MOUNTED_READ_ONLY.equals(externalStorageState));
-            menu.findItem(R.id.wallet_options_backup_wallet).setEnabled(Environment.MEDIA_MOUNTED.equals(externalStorageState));
-            menu.findItem(R.id.wallet_options_encrypt_keys).setTitle(
-                    wallet.isEncrypted() ? R.string.wallet_options_encrypt_keys_change : R.string.wallet_options_encrypt_keys_set);
-
-            boolean showFingerprintOption = fingerprintHelper != null && !fingerprintHelper.isFingerprintEnabled();
-            menu.findItem(R.id.wallet_options_enable_fingerprint).setVisible(showFingerprintOption);
         }
     }
 
@@ -928,8 +817,6 @@ public final class WalletActivity extends AbstractBindServiceActivity
             SweepWalletActivity.start(this, true);
         } else if (id == R.id.nav_network_monitor) {
             startActivity(new Intent(this, NetworkMonitorActivity.class));
-        } else if (id == R.id.nav_safety) {
-            openContextMenu(viewFakeForSafetySubmenu);
         } else if (id == R.id.nav_settings) {
             startActivity(new Intent(this, PreferenceActivity.class));
         } else if (id == R.id.nav_disconnect) {
@@ -937,22 +824,11 @@ public final class WalletActivity extends AbstractBindServiceActivity
         } else if (id == R.id.nav_report_issue) {
             handleReportIssue();
         }
-
-        viewDrawer.closeDrawer(GravityCompat.START);
         return true;
     }
 
-    @Override
-    public void onBackPressed() {
-        if (viewDrawer.isDrawerOpen(GravityCompat.START)) {
-            viewDrawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
     private void startUpholdActivity() {
-        startActivity(UpholdAccountActivity.createIntent(this, wallet));
+        startActivity(BuyAndSellLiquidUpholdActivity.Companion.createIntent(this));
     }
 
     //Dash Specific
@@ -1007,40 +883,12 @@ public final class WalletActivity extends AbstractBindServiceActivity
             return;
         }
 
-        int percentage = blockchainState.getPercentageSync();
-        if (blockchainState.getReplaying() && blockchainState.getPercentageSync() == 100) {
-            //This is to prevent showing 100% when using the Rescan blockchain function.
-            //The first few broadcasted blockchainStates are with percentage sync at 100%
-            percentage = 0;
-        }
-
-        ProgressBar syncProgressView = findViewById(R.id.sync_status_progress);
         if (blockchainState != null && blockchainState.syncFailed()) {
-            updateSyncPaneVisibility(R.id.sync_status_pane, true);
-            findViewById(R.id.sync_progress_pane).setVisibility(View.GONE);
             findViewById(R.id.sync_error_pane).setVisibility(View.VISIBLE);
             return;
         }
 
         updateSyncPaneVisibility(R.id.sync_error_pane, false);
-        updateSyncPaneVisibility(R.id.sync_progress_pane, true);
-        TextView syncStatusTitle = findViewById(R.id.sync_status_title);
-        TextView syncStatusMessage = findViewById(R.id.sync_status_message);
-        syncProgressView.setProgress(percentage);
-        TextView syncPercentageView = findViewById(R.id.sync_status_percentage);
-        syncPercentageView.setText(percentage + "%");
-
-        if (blockchainState.isSynced()) {
-            syncPercentageView.setTextColor(getResources().getColor(R.color.success_green));
-            syncStatusTitle.setText(R.string.sync_status_sync_title);
-            syncStatusMessage.setText(R.string.sync_status_sync_completed);
-            updateSyncPaneVisibility(R.id.sync_status_pane, false);
-        } else {
-            syncPercentageView.setTextColor(getResources().getColor(R.color.dash_gray));
-            updateSyncPaneVisibility(R.id.sync_status_pane, true);
-            syncStatusTitle.setText(R.string.sync_status_syncing_title);
-            syncStatusMessage.setText(R.string.sync_status_syncing_sub_title);
-        }
     }
 
     /**
@@ -1086,7 +934,7 @@ public final class WalletActivity extends AbstractBindServiceActivity
     private void setDefaultCurrency() {
         String countryCode = getCurrentCountry();
         log.info("Setting default currency:");
-        if(countryCode != null) {
+        if (countryCode != null) {
             try {
                 log.info("Local Country: " + countryCode);
                 Locale l = new Locale("", countryCode);
@@ -1120,9 +968,9 @@ public final class WalletActivity extends AbstractBindServiceActivity
     }
 
     private String getCurrentCountry() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N){
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             return LocaleList.getDefault().get(0).getCountry();
-        } else{
+        } else {
             return Locale.getDefault().getCountry();
         }
     }
@@ -1147,7 +995,7 @@ public final class WalletActivity extends AbstractBindServiceActivity
             if (config.wasUpgraded()) {
                 showFiatCurrencyChangeDetectedDialog(currentCurrencyCode, newCurrencyCode);
             } else {
-                if(CurrencyInfo.hasObsoleteCurrency(newCurrencyCode)) {
+                if (CurrencyInfo.hasObsoleteCurrency(newCurrencyCode)) {
                     log.info("found obsolete currency: " + newCurrencyCode);
                     newCurrencyCode = CurrencyInfo.getUpdatedCurrency(newCurrencyCode);
                 }
